backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy>=2.0.0
psycopg2-binary==2.9.9
geoalchemy2==0.15.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.2
redis==5.1.1
scikit-learn==1.5.2
statsmodels==0.14.2
numpy==2.1.1
shapely==2.0.5
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
pandas==2.2.2
requests==2.31.0
urllib3==2.2.0